+++ packages/agent-core/src/schema/index.ts
@@ -1,3 +1,4 @@
 // This file is meant to be imported from both sever and client side.
 
 export * from './events';
+export * from './todo';
