+++ packages/agent-core/src/tools/todo/index.ts
@@ -1,4 +1,4 @@
 export * from './todo-init';
 export * from './todo-update';
-export * from './types';
-export * from './todo-service';
+export * from '../../schema/todo';
+export * from '../../lib/todo';
